+++ message/command/tally-handler.ts
@@ -75,8 +75,13 @@ export default class TallyHandler {
                 .setTitle(`I could not ${TallyHandler.getBumpOrDump(isBump)}.`)
                 .setDescription(`${e.message}`);
         }
-
-        if (richEmbed) message.channel.send(richEmbed);
+        const sentMsg: any = await message.channel.send(richEmbed);
+        try {
+            await sentMsg.react('🔼');
+            await sentMsg.react('🔽')
+        } catch (e) {
+            logger.error(`error occured while attempting to react...`, e);
+        }
         CmdHelper.finalize(message);
     }